requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pylint==3.0.0
pytest==7.4.3
pytest-cov==4.1.0
requests-mock==1.11.0